+++ packages/vendure-plugin-stripe-subscription/src/api/types/stripe-invoice.ts
@@ -1,4 +1,4 @@
-import { Metadata } from './stripe.types';
+import { Metadata } from './stripe.common';
 
 export interface StripeInvoice {
   id: string;