+++ packages/vendure-plugin-invoices/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.0 (2023-10-24)
+
+- Updated vendure to 2.1.1
+
 # 1.1.0 (2023-08-14)
 
 - Allow specifying invoice template storage in DB by specifying DB engine: INVOICES_PLUGIN_DB_ENGINE=mysql ([#243](https://github.com/Pinelab-studio/pinelab-vendure-plugins/pull/243))