+++ packages/vendure-plugin-picqer/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.1.0 (2023-11-02)
+
+- Updated vendure to 2.1.1
+
+# 2.0.2
+
+- Try to fulfill again on order status changed webhooks, to support fulfilling of back orders.
+
+# 2.0.1
+
+- Push to Picqer even when fulfilling failed. Log order codes and manual steps needed when fulfillment fails.
+
+# 2.0.0
+
+- Complete orders on `orders.status_changed` webhooks, instead of `picklists.closed` hooks. ([#281](https://github.com/Pinelab-studio/pinelab-vendure-plugins/pull/281))
+
 # 1.0.13
 
 - Patch priority order of names when sending to Picqer: invoicename ?? deliveryname ?? customerFullname