+++ packages/vendure-plugin-popularity-scores/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.4.0 (2023-11-02)
+
+- Updated vendure to 2.1.1
+
+# 1.3.0(2023-10-30)
+
+- Add condition to check if a collection is empty as specified in (here)[https://github.com/Pinelab-studio/pinelab-vendure-plugins/issues/279]
+
 # 1.1.0
 
 - Remove unused select statement's while calculating product popularity scores to support Postgres